Extracted AuthenticationTokenProvider to separate file
