+++ src/services/data-targets/fiware-data-target.service.ts
@@ -1,5 +1,4 @@
-import { HttpService, Injectable, Logger, Inject, CACHE_MANAGER } from "@nestjs/common";
-import { Cache } from 'cache-manager'
+import { HttpService, Injectable, Logger } from "@nestjs/common";
 import { AxiosRequestConfig } from "axios";
 
 import { AuthorizationType } from "@enum/authorization-type.enum";
@@ -9,49 +8,7 @@ import { TransformedPayloadDto } from "@dto/kafka/transformed-payload.dto";
 import { DataTargetSendStatus } from "@interfaces/data-target-send-status.interface";
 import { FiwareDataTargetConfiguration } from "@interfaces/fiware-data-target-configuration.interface";
 import { BaseDataTargetService } from "@services/data-targets/base-data-target.service";
-
-type TokenEndpointResponse = {
-    access_token: string,
-    expires_in: number,
-}
-
-@Injectable()
-export class AuthenticationTokenProvider {
-
-    private readonly logger = new Logger(AuthenticationTokenProvider.name);
-
-    constructor(private httpService: HttpService, @Inject(CACHE_MANAGER) private cacheManager: Cache) {
-    }
-
-    async getToken(config: FiwareDataTargetConfiguration): Promise<string> {
-        const key = config.clientId + config.clientSecret;
-        const token = await this.cacheManager.get<string>(key)
-        if (token) {
-            this.logger.debug('Token found')
-            return token
-        } else {
-            try {
-                const encodedCredentials = Buffer.from(`${config.clientId}:${config.clientSecret}`).toString('base64');
-                const params = new URLSearchParams([['grant_type', 'client_credentials']])
-                const { data }: { data: TokenEndpointResponse } = await this.httpService.post(config.tokenEndpoint, params, {
-                    headers: {
-                        'Authorization': `Basic ${encodedCredentials}`,
-                        'Content-Type': 'application/x-www-form-urlencoded',
-                    },
-                }).toPromise()
-
-                const clockSkew = 30
-                const ttl = data.expires_in - clockSkew
-                this.logger.debug(`AuthenticationTokenProvider caching token for ${config.clientId} (expires in ${ttl} seconds)`)
-                this.cacheManager.set(key, data.access_token, { ttl })
-                return data.access_token
-            }
-            catch (err) {
-                this.logger.error(`AuthenticationTokenProvider got error ${err}`)
-            }
-        }
-    }
-}
+import { AuthenticationTokenProvider } from "../../helpers/fiware-token.helper";
 
 @Injectable()
 export class FiwareDataTargetService extends BaseDataTargetService {
